+++ common.h
@@ -47,8 +47,11 @@
 #include <pthread.h>
 
 #include<unordered_map>
-#include<vector>
-#include<string>
+#include <fstream>
+#include <string>
+#include <vector>
+#include <map>
+#include <set>
 using  namespace std;
